chore(model,api): move PXE from model to API
The idea of model as separate crate is to share data types
between different crates. PXE data types are only used in API.

This PR moves these data types from model to api crate.
In addition, it reduces dependency model crate from rpc crate.
