+++ crates/api-model/src/lib.rs
@@ -79,7 +79,6 @@ pub mod os;
 pub mod power_manager;
 pub mod power_shelf;
 pub mod predicted_machine_interface;
-pub mod pxe;
 pub mod rack;
 pub mod rack_firmware;
 pub mod rack_type;
