+++ crates/api-model/src/rpc_conv/mod.rs
@@ -51,7 +51,6 @@ pub mod operating_system_definition;
 pub mod os;
 pub mod power_manager;
 pub mod power_shelf;
-pub mod pxe;
 pub mod rack;
 pub mod rack_firmware;
 pub mod rack_type;
